requirements.txt
Flask
Authlib
google-cloud-retail
google-cloud-discoveryengine
gunicorn
python-dotenv